pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-auth>=2.38.0",
    "gspread>=6.2.0",
    "gunicorn>=23.0.0",
    "oauth2client>=4.1.3",
    "playwright>=1.51.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.40",
    "trafilatura>=2.0.0",
]
